revert: restore must not write stream state, and mine could have
The interagent reconciler added in #63 corrected a divergent stream from
activerestore. It should not have, for a reason that is structural rather
than a matter of taste.

Restore talks to the store over HTTP and takes no part in the Locker
protocol that serialises transactions against a stream. So a write from
there can land on top of a transaction committing on this node at that
moment, and because it wrote with force_rev - which checks nothing - it
would have done so silently, destroying a commit. That is exactly why
restore's remit was only ever to ADD documents this node is missing
(umids) rather than overwrite ones it has.

Worse, it did not honour the { _id, locked: true } markers introduced in
the same PR. A marker carries no _rev, so the tally skipped it as though
the node had said nothing - the precise defect that was fixed on the SPI
side and not carried across. It would therefore have voted on a sample
taken while a transaction held the stream on other nodes, which is when
overwriting is most dangerous.

It also never ran in production. SPI is triggered by the same event and
gets there 10-50x sooner, so the reconciler consistently found the stream
already repaired. The one demonstrated heal of a real laggard was SPI:
"SPI NOWINNER #2 ... sample incomplete" while the round was in flight,
then "SPI REWRITING #2 088067... @ 41-a9f4ef46 NOT 39-246bc890" once it
settled. Reconciling a stream belongs there, inside the transaction's own
lifecycle, where the locked markers are respected and it abstains rather
than guessing.

Removed: stream-resync.ts, the resync routing and its retry/attempt-cap,
and the tests for them. The umid-not-found path returns to marking the
document processed as it did before.

Kept: the 1200 error document from protocol/process.ts. It writes to the
error database, never to a stream, so it sits inside restore's existing
remit - and it is a durable record that this node disagreed about a
stream, which did not exist before and which would have found a
production laggard in minutes instead of hours. The interagent now
archives it and does nothing else with it.

New tests pin the rule directly: a position error is recorded without the
stream being touched, and restore does not poll peers for a revision to
write.
